CH: Moving logger into seperate file
