+++ packages/program-boilerplate/src/index.ts
@@ -1,10 +1,10 @@
 import * as express from 'express';
-import {createLogger, format, Logger, transports} from 'winston';
 
 import {meetCustomFieldRules, meetEventTriggerRules} from './conversion';
 import {rewardEmailQuery} from './queries';
 import Transaction from './transaction';
 import {triggerProgram} from './trigger';
+import {getLogger} from './logger';
 import * as types from './types';
 
 import {
@@ -47,6 +47,7 @@ export {
   getTriggerSchema,
   timeboxExpression,
   safeJsonata,
+  getLogger,
 };
 
 /**
@@ -87,33 +88,4 @@ export function webtask(program: Program = {}): express.Application {
   });
 
   return app;
-}
-
-// Lazily initialized logger instance
-let _logger: Logger;
-
-/**
- * Returns a logger for the programs to use instead of
- * console.log
- *
- * @param {string} logLevel The log level
- *
- * @return {Logger} The winston logger
- */
-export function getLogger(logLevel: string): Logger {
-  if (_logger) {
-    return _logger;
-  }
-
-  const logFormat = format.printf(({level, message}) => {
-    return `[${level.toUpperCase()}] ${message}`;
-  });
-
-  _logger = createLogger({
-    level: logLevel,
-    format: format.combine(logFormat),
-    transports: [new transports.Console()],
-  });
-
-  return _logger;
 }
